blocksize: Process requests in linear order
Mixing the blocksize filter with the streaming plugin is likely to
fail, because streaming does not handle read-modify-write. Still, the
idea that a plugin may expect data in linear order, rather than a
filter rewriting things into random-access order, is something worth
addressing - as such, the blocksize filter should not handle an
unaligned tail until after the aligned body of a large request.

Note that blocksize, cache, and cow all do block fragmentation and
bounce-buffer alignment, but only the blocksize filter was doing
random-access visiting of the tail before the body (conversely, the
blocksize filter is lazy and relies on serialized access to reuse a
single bounce buffer, while the other two malloc on demand and used
the bounce buffer more frequently).  It is likely that in the future,
we will migrate the best parts of the different block handling code
implementations into nbdkit proper, and allow plugins/filters the
ability to report their preferred alignment and block sizes, at which
point the blocksize filter would be a much thinner task of just
altering the numbers that would otherwise be reported by the
plugin. For now, getting the logic consistent between the three
filters will make that future refactoring easier.

Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit ede021a6024b36b50a3347d89b75a03486538306)