server/crypto: Don't need to explicitly malloc the session pointer.
gnutls_session_t is a pointer (not a larger object), so it doesn't
have to be explicitly allocated.  In addition we can store the pointer
directly in the ‘void *’ of the connection (instead of storing a
pointer to a pointer).  The old code wasn't wrong but did an
unnecessary and complex extra allocation.