cache, cow: Round size down
The blocksize filter already rounds sizes down, to avoid having to
special-case a partial action at the end of an unaligned file.  But
our current behavior for cache and cow is not similarly careful, and
can end up causing errors by passing a too-large count to the
plugin. The bulk of our plugins are somewhat safe, in that they either
fail the pread (for example, the file plugin gets a short read due to
EOF) or populate the tail of the buffer with other data (which the
client will never see, because we at least clamp user requests), so
that the cache is only populated with untainted data. But it's hard to
argue whether there might be another plugin out there that might abort
because we violated our documented constraints, or which might leave
the tail of the buffer unpopulated on a short read while still
reporting success. Thus, I cannot rule out that our cache file might
contain uninitialized data from the heap, or even leak it back out to
the file system through the plugin or the NBD client.

The quickest fix is to just always round down, and then do a better
job of documenting how the user can round up as needed.

It would be nice if we had parameter validation of each filter's call
into the next layer, rather than only at the client's call into the
first filter (such checking would have flagged the bug we fixed here);
but that's a more invasive effort for a later day. Similarly, it would
be nice if we had 'nbdkit --dump-filter-config --filter=cache' for
letting the user programmatically learn the block size in effect, but
also an idea for a later day.

Signed-off-by: Eric Blake <eblake@redhat.com>