noparallel: Implement new filter
Similar to the existing fua, nocache, nozero and noextents filters,
add a filter to make it easy to override the plugin's choice of
parallel processing.  Several possible uses:

- facilitate timing tests of whether a plugin's parallel support
and/or client's use of multi-conn increases throughput

- enable a quick workaround of any plugin that was mistakenly compiled
with a claim of parallel support but with insufficient locking

- serve a parallel plugin to a client that batches up requests but is
not prepared to handle out-of-order replies (however, this does not
necessarily prevent deadlock if the client batches a large write
immediately behind a large read before reading the server's response
to the read)

Update test-eflags (showing the effect on the multi-conn flag) and
test-parallel-file (showing the effect on preventing out-of-order
replies).

In implementing this, I wonder if we should have a thread model
in between serialize-requests and parallel. Right now, if I use
--filter=delay rdelay=2 wdelay=3, coupled with a client that starts a
write followed by a read, the behaviors are:

Parallel (filter not used), latency 3, out-of-order replies:
0   1   2   3   4   5
+---+---+---+---+---+
=>write request
 =>read request
         <=read reply
	    <=write reply

Serialize-requests (--filter=noparallel), latency 5, in-order replies:
0   1   2   3   4   5
+---+---+---+---+---+
=>write request
	    <=write reply
 ============>read request
                     <=read reply

But merely guaranteeing in-order completion falls in between:

New mode, latency 3, in-order replies:
0   1   2   3   4   5
+---+---+---+---+---+
=>write request
	    <=write reply
 =>read request
         <=====read reply

Since nbdkit is already farming out requests to different threads, it
could keep a list of which thread was fired off in which order, then
use pthread_cond_broadcast and friends to block a thread from sending
the reply to the client until all other threads with earlier
sequencing have completed first. But adding such a mode is tricky -
our existing modes are API in nbdkit-common.h, so we can't renumber
NBDKIT_THREAD_MODEL_PARALLEL, yet adding a new mode 5 that falls in
between 3 and 4 changes all our existing code that has so far relied
on descending values being stricter. At the same time, such a mode HAS
to be done in nbdkit proper - by the time a filter's callbacks are
reached within separate threads, scheduling races means the filter has
no reliable way of knowing which of the two threads represents the
earlier client request.

Anyways, if nbdkit adds such a threading model, noparallel would be
the obvious filter to allow easy use of that model with any
pre-existing plugin :)

Signed-off-by: Eric Blake <eblake@redhat.com>