nocache: Implement new filter
Similar to the existing fua, nozero and noextents filters, add a
filter to make it easy to override the basic caching functionality, in
part to facilitate timing tests of whether a plugin's cache
implementation is worthwhile.

A worthwhile test to add to the testsuite would connect the log filter
both before and after the nocache filter, to prove how caching
requests are altered.  However, until we use libnbd in the nbdkit
testsuite, there isn't really a clean way to trigger NBD_CMD_CACHE
calls to test this (qemu-nbd can't do it). But I can at least exploit
test-eflags to use 'qemu-nbd --list' to demonstrate that the
advertised flag changes, with a little tweak to allow the test to run
filters with arguments.

Signed-off-by: Eric Blake <eblake@redhat.com>