test-eflags: Fix multi-conn check
In the sh plugin, the boolean 'can_FOO' functions work by exit status
alone, rather than by echoing to stdout (only the tri-state functions
need to do that).  The exit status happened to be correct in this
script, but had we tried to 'echo 3' instead of 'exit 3', the mistake
would be more obvious.  Typo'd in commit d5eeae41.

Signed-off-by: Eric Blake <eblake@redhat.com>