filters: Use .thread_model callback, not THREAD_MODEL define
We already documented that filters should default to being safe with
THREAD_MODEL_PARALLEL; and while not all of our filters started out
that way, we have over time improved things so that all but the xz
filter currently comply. Furthermore, the next patch wants to add a
filter that can dynamically tighten the thread model based on config
options.  Since we already select the strictest model, it's easier to
state that a filter doesn't have to redundantly declare a thread model
except when it is tightening things via a callback.  This is an ABI
change, but since all filters are in-tree and we already have strict
checks for same-version matching, this is a safe change.

Signed-off-by: Eric Blake <eblake@redhat.com>