server: Defer thread model selection until after config
The backend code already uses a function callback at each level of the
filter/plugin chain to select the most restrictive model. In the
future, we want this model to be selected dynamically based on
configuration options.  For example, several of our language bindings
are currently hard-coded at SERIALIZE_ALL_REQUESTS [due to
difficulties in writing multi-thread-safe bindings], but a script in
that language may wish to further restrict things to
SERIALIZE_CONNECTIONS.  Also, having a filter that can force a
stricter model on the fly makes sense.

To do this, we need to defer the initialization of our choice of locks
based on threading model until after .config_complete, rather than
immediately after .load. Add some assertions to show this move is
safe.

Signed-off-by: Eric Blake <eblake@redhat.com>