plugins: Add .thread_model callback
Plugins cannot get rid of #define THREAD_MODEL (doing so would break
API/ABI compatibility); however, we CAN add an optional callback for
allowing a runtime reduction of the thread model. This can be
especially useful for language bindings, as in the next patch, where
the C glue code for the language has to cater to the maximum
parallelism possible for that language, but where an individual plugin
may still want to be stricter.

Conversely, it is intentional that the .thread_model callback can
request to be more parallel than THREAD_MODEL, and such requests are
silently ignored. This allows us to release language bindings that
expose only the callback form in other languages, while still defining
THREAD_MODEL to match the restrictions of the glue code; if we can
later figure out how to make the glue code more parallel, the script
plugin can automatically benefit from that when run against newer
nbdkit.

Signed-off-by: Eric Blake <eblake@redhat.com>