protocol: Fix base:allocation replies when req_one is not set.
We never tested this path because qemu always sets the req_one flag,
but what we sent back was totally broken.  The length field was not
set at all, and the count of blocks could be off by one in some cases.
These problems were revealed while writing libnbd tests.