nbd: Add retry=N parameter
A common command line use case for the nbd plugin is
 ( sock=`mktemp -u`
   nbdkit --exit-with-parent nbd socket=$sock &
   exec /path/to/server -k $sock )

As long as the first client to nbdkit doesn't try to connect until
after /path/to/server has created $sock, we are fine. But on a heavily
loaded system, it would be nice to give nbdkit some leeway to try
again to connect to the socket if a client connects quickly to nbdkit.

This becomes even more useful in the next patch when adding an option
for the nbd plugin to connect during .config_complete rather than
during .open, as the window of time for the external server to make
the socket available is no longer dependent on when nbdkit's first
client arrives.

Signed-off-by: Eric Blake <eblake@redhat.com>