data, memory: Advertise flush support
The data and memory plugins have no work to do on a flush (after all,
everything is already in memory, which is all the more persistent any
writes will ever be). However, a client wanting to perform flushes for
safety may refuse to connect to a server that cannot do flushes.  As
such, it is better to follow the lead of the null plugin in
advertising explicit flush and FUA support, even if the support is a
no-op.  (This is similar to how the kernel permits fsync() on tmpfs,
rather than failing it, even though tmpfs has nothing to sync).  This
requires upgrading the two plugins to use API version 2.

The other memory-only plugins either stick to read-only use (pattern,
random, zero), or advertise write support but don't actually permit
writes (full), so it is less of a problem if a client refuses to
connect to those plugins when flush support is not advertised.

Signed-off-by: Eric Blake <eblake@redhat.com>