nbd: Check for libnbd
This patch merely adds the framework to compile the nbd plugin either
as stanadlone (the version frozen in time to this commit) or as a
client of the brand-new libnbd (a later patch will actually enable
that part).  Since libnbd does not yet have a stable API, falling back
to the standalone version makes sense for a while longer.

The configure check requires at least 0.1.3; at the moment, Fedora 29
has access to pre-built 0.1.4).

Signed-off-by: Eric Blake <eblake@redhat.com>