nbd: Use libnbd 0.1.3+
This conversion should be feature compatible with the standalone nbd
code. Note that the use of libnbd makes the binary for this particular
plugin fall under an LGPLv2+ license rather than BSD; but the source
code in nbd.c remains BSD.

A lot of code simply disappears, now that I'm no longer directly
utilizing the NBD protocol files but relying on libnbd.  Coordination
between threads from nbdkit and a central state machine thread is
gated by a lock around the pending transition list, using the same
trick as before of using a semaphore to wake up the right thread after
the server's reply, regardless of out-of-order handling from the
server.

Benchmark-wise, using a similar setup as in commit e897ed70 (the only
change being a larger export to track changes to libnbd
threaded-reads-and-writes in the meantime), I see a slight slowdown:

Pre-patch, the runs averaged  16.259s, 1.03E+10 bits/s
Post-patch, the runs averaged 17.585s, 9.54E+09 bits/s

A possible explanation for the 8% degredation is that pre-patch, one
nbdkit thread can be write()ing the next request to the server at the
same time the reader thread is read()ing the previous reply (using
blocking I/O); post-patch, libnbd's current structure handles all I/O
under a lock such that there is no simultaneous bidirectional traffic
(but I/O is non-blocking) and defers sending a request until the state
machine is not in the middle of handling a reply. There are
discussions about adding future libnbd APIs to address this
limitation, and rewriting this use case on top of any proposed
dedicated writer thread will help prove if it is worthwhile.

The trickiest part (for me) was the fact that since the state machine
loop is in a separate thread from the initial requests, the loop is
often blocked on just POLLIN for the state machine fd.  My initial
attempt tried to grab the transaction lock before calling nbd_aio_FOO
to create the cookie, and generally worked where the server's reply
would then wake up the reader. But that was even slower (due to a
larger lock section), and also hit a problem where large NBD_CMD_WRITE
would block on writes, which the poll() was not expecting (best seen
when running tests/test-nozero.sh with LIBNBD_DEBUG=1).  I tried
blindly setting POLLOUT as a valid reason to break the poll without
regard to the current state, but that burned a lot of CPU as it fired
even when there was no progress.  So I finally settled on using a
pipe-to-self and poll()ing on two fds at once (the libnbd state
machine and my pipe) to know when the loop needed to start processing
because another thread started a command, where the pipe-to-self also
makes it possible to not need to hold the transaction lock while
generating a cookie.

Signed-off-by: Eric Blake <eblake@redhat.com>