crypto: Tweak handling of SEND_MORE
In the recent commit 3842a080 to add SEND_MORE support, I blindly
implemented the tls code as:
 if (SEND_MORE) {
   cork
   send
 } else {
   send
   uncork
 }
because it showed improvements for my test case of aio-parallel-load
from libnbd. But that was when the test was hard-coded to 64k
transactions; in later testing, I found that it reliably pessimizes
performance for 2M transactions.  Why?

It turns out that when gnutls is corked, it memcpy()s the entire data
into realloc()d memory, and doesn't encrypt or send anything over the
wire until uncorking. If you uncork before the point at which TCP
would have fragmented across multiple segments, it's a win; but the
larger the packet gets, the more time is spent moving data around in
memory, and the more time you've lost in which you could have
otherwise sent the short header in parallel with encrypting the main
payload that was going to be fragmented anyway.

After more measurements on my machine, I saw the following IOPS:
  packet size:	512		32k		64k		2m
No cork [1]:	325,209.3	187,598.9	116,304.2	4,856.7
full cork [2]:	461,083.8	198,009.3	118,072.2	4,595.1
32k cork:	456,365.5	196,586.5	117,772.4	4,779.1
64k cork [3]:	450,443.6	194,505.9	117,226.6	4,822.2
128k cork:	455,049.4	197,991.4	117,157.9	4,953.7

[1] Using the code prior to 3842a080
[2] Using the code from 3842a080
[3] The value chosen for this patch

Note that the lower IOPS for larger NBD_CMD_READ sizes in relation to
smaller sizes is misleading, the connection is actually achieving a
higher bits-per-second throughput (less packet overhead); the real
goal is increasing IOPS within a column. The table also shows that
there is some measurement noise (the line for 64k almost looks worse
than for 128k, but both seem better than no corking at all); and also
backs up my claim that for the 2m column, not corking for large
packets really is the right way to go.  My choice for 64k as my
heuristic cutoff is based on that being the theoretical maximum TCP
segment size.

Signed-off-by: Eric Blake <eblake@redhat.com>