Introduce cacheextents filter
This filter caches the last result of the extents() call and offers a nice
speed-up for clients that only support req_one=1 in combination with plugins
like vddk, which has no overhead for returning information for multiple extents
in one call, but that call is very time-consuming.

Quick test showed that on a fast connection and a sparsely allocated 16G disk
with a OS installed `qemu-img map` runs 16s instead of 33s (out of which it
takes 5s to the first extents request).  For 100G disk with no data on it, that
is one hole extent spanning the whole disk (where there is no space for
improvement) this does not add any noticeable overhead.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <ce08f27c1bb7cc2c6c590ce017f23c16a8ded43b.1560242961.git.mkletzan@redhat.com>
[eblake: minor tweaks]
Signed-off-by: Eric Blake <eblake@redhat.com>

For RHEL 8.1: Small modifications were required to support the
ACQUIRE_LOCK_FOR_CURRENT_SCOPE macro.

(cherry picked from commit e007f066008407efac5bd888cdcbd09a4deb7052)