captive: Support $uri in --run
The existing --run '$nbd' outputs an older form that differs between
libguestfs and qemu, and which is not always a valid URI.  For
historical compatibility, we probably can't change that; but we can
instead add a new '$uri' that outputs a valid URI.  Note that the
libguestfs '$nbd' TCP form is already a valid URI (at least, as long
as the user isn't exploiting $TMPDIR or -U /odd/#path), but that
libguestfs still needs to be taught about the nbd+unix:// scheme, so
for there, you are still better off using '$nbd'; but for qemu, using
'$uri' already works since qemu v1.3.0 in 2012.

Note that the NBD project has not actually yet posted a link for the
preferred URI syntax; once Rich is happy with his work on that
project, we can touch this up to link to that page.

Reported-by: Martin Kletzander <mkletzan@redhat.com>
Signed-off-by: Eric Blake <eblake@redhat.com>