nozero: Work around gcc 9 size pessimization
gcc 9 decided [1] that implicitly-zero-initialized static const
buffers belong in .rodata rather than .bss (for the extra safety to
ensure you can't accidentally modify the buffer contents).  However,
this inflates our binary size.  Avoiding 'const' puts things back in
.bss, and it's still easy enough to prove (with reduced scope in
nozero.c to match what was already in plugins.c) that the buffer is
not modified in spite of dropping the keyword that would let the
compiler prove it on our behalf.

[1] https://gcc.gnu.org/ml/gcc/2019-04/msg00034.html

Reported-by: Thomas Weißschuh <thomas@weissschuh.net>
Signed-off-by: Eric Blake <eblake@redhat.com>