delay: Avoid numeric overflow
Attempting delay-read=1000 results in no delay whatsoever: 1000
seconds, scaled to milliseconds, stored in int, then subjected to

      .tv_nsec = (ms * 1000000) % 1000000000

results in .tv_nsec being set to -727379968 thanks to 32-bit overflow
of ms * 1000000, but that in turn results in instant EINVAL failure of
nanosleep().

Fix it by diagnosing failure to fit in an int during config, and avoid
math that overflows during delay().  Forbid negative delays while at it.

Signed-off-by: Eric Blake <eblake@redhat.com>